let initial vDOM render be diff against existing DOM node
Experimental.

If the parent node handed to the virtual DOM on initialization has
contents, use that contents to reconstruct an initial state of the
vDOM, and immediately schedule an update to render the node produced by
the elm program. In this way if the initial page contents is close to
the output of the program, the first update will be a small
“progressive” diff.